ai_image_generator/requirements.txt
openai==0.28.1
requests==2.31.0
Pillow==10.0.1
python-dotenv==1.0.0 